% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsmatr.R
\name{ggsmatr}
\alias{ggsmatr}
\title{ggsmatr 'ggplot2' based SMATR plots}
\usage{
ggsmatr(data, groups , xvar , yvar, sma.fit)
}
\arguments{
\item{data}{a dataframe.}

\item{groups}{group variable.}

\item{xvar}{x variable for drawing.}

\item{yvar}{y variable for drawing.}

\item{sma.fit}{an sma or ma object fitted in SMATR package.}
}
\value{
ggplot based plot of sma.
}
\description{
Create a scatter plot based on the coefficients from (Standardised) Major Axis Estimation fit.
}
\examples{
datafile <- system.file("iris.csv", package = "ggsmatr")
df.iris <- read.csv(datafile, encoding = "UTF-8")

library(ggsmatr)
library(ggplot2)
library(smatr)
fit = sma(Sepal.Length ~ Sepal.Width + Species, 
          data = df.iris,shift=TRUE, elev.test=TRUE)
#
ggsmatr(data =  df.iris, groups = "Species", 
        xvar =  "Sepal.Width", yvar = "Sepal.Length", 
        sma.fit =  fit) + 
theme(legend.position = "top", legend.title=element_blank())+ 
ylab("Sepal.Length")+ 
xlab("Sepal.Width")+ 
theme(legend.position = "top")


}
