% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsmatr.R
\encoding{UTF-8}
\name{ggsmatr}
\alias{ggsmatr}
\title{'ggplot2' Based SMATR Plots}
\usage{
ggsmatr(data, groups , xvar , yvar, sma.fit)
}
\arguments{
\item{data}{a dataframe.}

\item{groups}{group variable.}

\item{xvar}{x variable for drawing.}

\item{yvar}{y variable for drawing.}

\item{sma.fit}{an sma or ma object fitted in SMATR package.}
}
\value{
ggplot based plot of sma.
}
\description{
Create a scatter plot based on the coefficients from (Standardised) Major Axis Estimation fit.
}
\examples{
#data(iris)
#library(smatr)
#fit <- sma(Petal.Width ~ Petal.Length+ Species, data=iris)
#ggsmatr(data =  iris, groups = "Species", xvar =  "Petal.Length", yvar =
#"Petal.Width", sma.fit =  fit)

#Do not run, just examples.
#ggsmatr(data =  df1, groups = "Site", xvar =  "Point.count", yvar =
#"Mist.nets", sma.fit =  fit)
#ggsmatr(data =  evdata, groups = "site.method", xvar =  "Species.richness",
#yvar ="EvennesObserved", sma.fit =  fit)
#Plot is saved as ggp object, and can be edited using ggplot parameters.
#ggp + xlab("x axis name") + ylab("y axis name") +theme(legend.position = "right",
#legend.title=element_blank())
}
